# Use segment local hit and miss counter.