--- a/cache.go
+++ b/cache.go
@@ -9,10 +9,8 @@ import (
 )
 
 type Cache struct {
-	locks     [256]sync.Mutex
-	segments  [256]segment
-	hitCount  int64
-	missCount int64
+	locks    [256]sync.Mutex
+	segments [256]segment
 }
 
 func hashFunc(data []byte) uint64 {
@@ -53,11 +51,6 @@ func (cache *Cache) Get(key []byte) (value []byte, err error) {
 	cache.locks[segId].Lock()
 	value, _, err = cache.segments[segId].get(key, hashVal)
 	cache.locks[segId].Unlock()
-	if err == nil {
-		atomic.AddInt64(&cache.hitCount, 1)
-	} else {
-		atomic.AddInt64(&cache.missCount, 1)
-	}
 	return
 }
 
@@ -68,11 +61,6 @@ func (cache *Cache) GetWithExpiration(key []byte) (value []byte, expireAt uint32
 	cache.locks[segId].Lock()
 	value, expireAt, err = cache.segments[segId].get(key, hashVal)
 	cache.locks[segId].Unlock()
-	if err == nil {
-		atomic.AddInt64(&cache.hitCount, 1)
-	} else {
-		atomic.AddInt64(&cache.missCount, 1)
-	}
 	return
 }
 
@@ -153,20 +141,31 @@ func (cache *Cache) AverageAccessTime() int64 {
 	}
 }
 
-func (cache *Cache) HitCount() int64 {
-	return atomic.LoadInt64(&cache.hitCount)
+func (cache *Cache) HitCount() (count int64) {
+	for i := range cache.segments {
+		count += atomic.LoadInt64(&cache.segments[i].hitCount)
+	}
+	return
+}
+
+func (cache *Cache) MissCount() (count int64) {
+	for i := range cache.segments {
+		count += atomic.LoadInt64(&cache.segments[i].missCount)
+	}
+	return
 }
 
 func (cache *Cache) LookupCount() int64 {
-	return atomic.LoadInt64(&cache.hitCount) + atomic.LoadInt64(&cache.missCount)
+	return cache.HitCount() + cache.MissCount()
 }
 
 func (cache *Cache) HitRate() float64 {
-	lookupCount := cache.LookupCount()
+	hitCount, missCount := cache.HitCount(), cache.MissCount()
+	lookupCount := hitCount + missCount
 	if lookupCount == 0 {
 		return 0
 	} else {
-		return float64(cache.HitCount()) / float64(lookupCount)
+		return float64(hitCount) / float64(lookupCount)
 	}
 }
 
@@ -184,13 +183,9 @@ func (cache *Cache) Clear() {
 		cache.segments[i] = newSeg
 		cache.locks[i].Unlock()
 	}
-	atomic.StoreInt64(&cache.hitCount, 0)
-	atomic.StoreInt64(&cache.missCount, 0)
 }
 
 func (cache *Cache) ResetStatistics() {
-	atomic.StoreInt64(&cache.hitCount, 0)
-	atomic.StoreInt64(&cache.missCount, 0)
 	for i := 0; i < 256; i++ {
 		cache.locks[i].Lock()
 		cache.segments[i].resetStatistics()