Use a global statistic counter have some scalability problem at high contention. By relax the consistency restrict, use a pre-segment local counter will have better performance.

Use the following naive benchmark for a simple evaluation (running on my old 8-core i7 system with Go1.10rc1).
```go
func BenchmarkCacheGetConcurrent(b *testing.B) {
	cache := NewCache(256 * 1024 * 1024)
	var key [8]byte
	const size = 100000
	for i := 0; i < size; i++ {
		binary.LittleEndian.PutUint64(key[:], uint64(i))
		cache.Set(key[:], make([]byte, 8), 0)
	}
	b.ResetTimer()

	b.RunParallel(func(pb *testing.PB) {
		var key [8]byte
		rnd := rand.New(rand.NewSource(rand.Int63()))
		for pb.Next() {
			binary.LittleEndian.PutUint64(key[:], uint64(rnd.Intn(size)))
			cache.Get(key[:])
		}
	})
}
```

```
go1.10rc1 test -run=nonthingplease -bench BenchmarkCacheGetConcurrent  -benchtime 5s -count 10
````

master:
```
BenchmarkCacheGetConcurrent-8           100000000               63.0 ns/op
BenchmarkCacheGetConcurrent-8           100000000               61.6 ns/op
BenchmarkCacheGetConcurrent-8           100000000               62.2 ns/op
BenchmarkCacheGetConcurrent-8           100000000               63.7 ns/op
BenchmarkCacheGetConcurrent-8           100000000               64.9 ns/op
BenchmarkCacheGetConcurrent-8           100000000               65.7 ns/op
BenchmarkCacheGetConcurrent-8           100000000               68.1 ns/op
BenchmarkCacheGetConcurrent-8           100000000               66.4 ns/op
BenchmarkCacheGetConcurrent-8           100000000               66.8 ns/op
BenchmarkCacheGetConcurrent-8           100000000               66.3 ns/op
```

this PR:
```
BenchmarkCacheGetConcurrent-8           100000000               58.1 ns/op
BenchmarkCacheGetConcurrent-8           100000000               56.0 ns/op
BenchmarkCacheGetConcurrent-8           100000000               57.4 ns/op
BenchmarkCacheGetConcurrent-8           100000000               62.1 ns/op
BenchmarkCacheGetConcurrent-8           100000000               62.9 ns/op
BenchmarkCacheGetConcurrent-8           100000000               62.1 ns/op
BenchmarkCacheGetConcurrent-8           100000000               61.6 ns/op
BenchmarkCacheGetConcurrent-8           100000000               63.3 ns/op
BenchmarkCacheGetConcurrent-8           100000000               62.8 ns/op
BenchmarkCacheGetConcurrent-8           100000000               63.8 ns/op
```