@coocood 2018-02-06
Nice optimization!

Thank you!